Ajustes adicionais para boleto
