+++ Model/Payment/AbstractMethod.php
@@ -1,0 +1,223 @@
+<?php
+
+namespace Vindi\Payment\Model\Payment;
+
+use Magento\Framework\DataObject;
+
+abstract class AbstractMethod extends \Magento\Payment\Model\Method\AbstractMethod
+{
+    /**
+     * @var \Vindi\Payment\Model\Payment\Api
+     */
+    protected $api;
+
+    /**
+     * @var \Magento\Sales\Model\Service\InvoiceService
+     */
+    protected $invoiceService;
+
+    /**
+     * @var Customer
+     */
+    protected $customer;
+
+    /**
+     * @var Product
+     */
+    protected $product;
+
+    /**
+     * @var Bill
+     */
+    protected $bill;
+
+    /**
+     * @var Profile
+     */
+    protected $profile;
+
+    /**
+     * @var PaymentMethod
+     */
+    protected $paymentMethod;
+
+    /**
+     * @var \Psr\Log\LoggerInterface
+     */
+    protected $psrLogger;
+
+    /**
+     * @var \Magento\Framework\Stdlib\DateTime\TimezoneInterface
+     */
+    protected $date;
+
+    public function __construct(
+        \Magento\Framework\Model\Context $context,
+        \Magento\Framework\Registry $registry,
+        \Magento\Framework\Api\ExtensionAttributesFactory $extensionFactory,
+        \Magento\Framework\Api\AttributeValueFactory $customAttributeFactory,
+        \Magento\Payment\Helper\Data $paymentData,
+        \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig,
+        \Magento\Payment\Model\Method\Logger $logger,
+        \Vindi\Payment\Model\Payment\Api $api,
+        \Magento\Sales\Model\Service\InvoiceService $invoiceService,
+        Customer $customer,
+        Product $product,
+        Bill $bill,
+        Profile $profile,
+        PaymentMethod $paymentMethod,
+        \Psr\Log\LoggerInterface $psrLogger,
+        \Magento\Framework\Stdlib\DateTime\TimezoneInterface $date,
+        \Magento\Framework\Model\ResourceModel\AbstractResource $resource = null,
+        \Magento\Framework\Data\Collection\AbstractDb $resourceCollection = null,
+        array $data = []
+    ) {
+        parent::__construct(
+            $context,
+            $registry,
+            $extensionFactory,
+            $customAttributeFactory,
+            $paymentData,
+            $scopeConfig,
+            $logger,
+            $resource,
+            $resourceCollection,
+            $data
+        );
+
+        $this->api = $api;
+        $this->invoiceService = $invoiceService;
+        $this->customer = $customer;
+        $this->product = $product;
+        $this->bill = $bill;
+        $this->profile = $profile;
+        $this->paymentMethod = $paymentMethod;
+        $this->psrLogger = $psrLogger;
+        $this->date = $date;
+    }
+
+    /**
+     * @return string
+     */
+    abstract protected function getPaymentMethodCode();
+
+    /**
+     * Check whether payment method can be used
+     *
+     * @param \Magento\Quote\Api\Data\CartInterface|null $quote
+     * @return bool
+     */
+    public function isAvailable(\Magento\Quote\Api\Data\CartInterface $quote = null)
+    {
+        return parent::isAvailable($quote);
+    }
+
+    /**
+     * Assign data to info model instance
+     *
+     * @param mixed $data
+     *
+     * @return $this
+     */
+    public function assignData(DataObject $data)
+    {
+        parent::assignData($data);
+        return $this;
+    }
+
+    /**
+     * Validate payment method information object
+     *
+     * @return $this
+     */
+    public function validate()
+    {
+        parent::validate();
+        return $this;
+    }
+
+    /**
+     * Authorize payment abstract method
+     *
+     * @param \Magento\Framework\DataObject|InfoInterface $payment
+     * @param float $amount
+     * @throws \Magento\Framework\Exception\LocalizedException
+     * @return $this|string
+     */
+    public function authorize(\Magento\Payment\Model\InfoInterface $payment, $amount)
+    {
+        parent::authorize($payment, $amount);
+        $this->processPayment($payment, $amount);
+    }
+
+    /**
+     * Capture payment abstract method
+     *
+     * @param \Magento\Framework\DataObject|InfoInterface $payment
+     * @param float $amount
+     * @throws \Magento\Framework\Exception\LocalizedException
+     * @return $this|string
+     */
+    public function capture(\Magento\Payment\Model\InfoInterface $payment, $amount)
+    {
+        parent::capture($payment, $amount);
+        $this->processPayment($payment, $amount);
+    }
+
+    /**
+     * @param \Magento\Framework\DataObject|InfoInterface $payment
+     * @param float $amount
+     * @throws \Magento\Framework\Exception\LocalizedException
+     * @return $this|string
+     */
+    protected function processPayment(\Magento\Payment\Model\InfoInterface $payment, $amount)
+    {
+        /** @var \Magento\Sales\Model\Order $order */
+        $order = $payment->getOrder();
+        $customerId = $this->customer->findOrCreate($order);
+        $productList = $this->product->findOrCreateProducts($order);
+
+        $body = [
+            'customer_id' => $customerId,
+            'payment_method_code' => $this->getPaymentMethodCode(),
+            'bill_items' => $productList
+        ];
+
+        if ($body['payment_method_code'] === PaymentMethod::CREDIT_CARD) {
+            $paymentProfile = $this->profile->create($payment, $customerId, $this->getPaymentMethodCode());
+            $body['payment_profile'] = ['id' => $paymentProfile['payment_profile']['id']];
+        }
+
+        if ($installments = $payment->getAdditionalInformation('installments')) {
+            $body['installments'] = (int)$installments;
+        }
+
+        if ($bill = $this->bill->create($body)) {
+            if ($body['payment_method_code'] === PaymentMethod::BANK_SLIP) {
+                $payment->setAdditionalInformation('print_url', $bill['charges'][0]['print_url']);
+                $payment->setAdditionalInformation('due_at', $bill['charges'][0]['due_at']);
+            }
+
+            if (
+                $body['payment_method_code'] === PaymentMethod::BANK_SLIP
+                || $body['payment_method_code'] === PaymentMethod::DEBIT_CARD
+                || $bill['status'] === Bill::PAID_STATUS
+                || $bill['status'] === Bill::REVIEW_STATUS
+                || $bill['charges'][0]['status'] === Bill::FRAUD_REVIEW_STATUS
+            ) {
+                $order->setVindiBillId($bill['id']);
+                return $bill['id'];
+            }
+            $this->bill->delete($bill['id']);
+        }
+
+        $this->psrLogger->error(__(sprintf('Error on order payment %d.', $order->getId())));
+        $message = __('There has been a payment confirmation error. Verify data and try again');
+        $order->setState(\Magento\Sales\Model\Order::STATE_CANCELED)
+            ->setStatus($order->getConfig()->getStateDefaultStatus(\Magento\Sales\Model\Order::STATE_CANCELED))
+            ->addStatusHistoryComment($message->getText());
+        throw new \Magento\Framework\Exception\LocalizedException($message);
+
+        return $this;
+    }
+}
