+++ Model/Payment/BankSlip.php
@@ -3,10 +3,9 @@
 namespace Vindi\Payment\Model\Payment;
 
 use Magento\Framework\DataObject;
-use Magento\Sales\Model\Order;
 use Vindi\Payment\Block\Info\BankSlip as InfoBlock;
 
-class BankSlip extends \Magento\Payment\Model\Method\AbstractMethod
+class BankSlip extends \Vindi\Payment\Model\Payment\AbstractMethod
 {
     const CODE = 'vindi_bankslip';
 
@@ -69,65 +68,6 @@ class BankSlip extends \Magento\Payment\Model\Method\AbstractMethod
      */
     protected $_canSaveCc = false;
 
-    protected $_invoiceService;
-    protected $api;
-    protected $order;
-
-    public function __construct(
-        \Magento\Framework\Model\Context $context,
-        \Vindi\Payment\Model\Payment\Api $api,
-        Customer $customer,
-        Product $product,
-        Bill $bill,
-        Profile $profile,
-        \Psr\Log\LoggerInterface $psrLogger,
-        \Magento\Framework\Stdlib\DateTime\TimezoneInterface $date,
-        PaymentMethod $paymentMethod,
-        \Magento\Sales\Model\Service\InvoiceService $invoiceService,
-        \Magento\Framework\Registry $registry,
-        \Magento\Framework\Api\ExtensionAttributesFactory $extensionFactory,
-        \Magento\Framework\Api\AttributeValueFactory $customAttributeFactory,
-        \Magento\Payment\Helper\Data $paymentData,
-        \Magento\Framework\App\Config\ScopeConfigInterface $scopeConfig,
-        \Magento\Payment\Model\Method\Logger $logger,
-        \Magento\Framework\Model\ResourceModel\AbstractResource $resource = null,
-        \Magento\Framework\Data\Collection\AbstractDb $resourceCollection = null,
-        array $data = []
-    ) {
-
-        $this->_logger = $logger;
-        parent::__construct(
-            $context,
-            $registry,
-            $extensionFactory,
-            $customAttributeFactory,
-            $paymentData,
-            $scopeConfig,
-            $logger,
-            $resource,
-            $resourceCollection,
-            $data
-        );
-
-        $this->api = $api;
-        $this->_invoiceService = $invoiceService;
-        $this->customer = $customer;
-        $this->product = $product;
-        $this->bill = $bill;
-        $this->paymentMethod = $paymentMethod;
-        $this->date = $date;
-        $this->psrLogger = $psrLogger;
-        $this->profile = $profile;
-    }
-
-
-    public function isAvailable(
-        \Magento\Quote\Api\Data\CartInterface $quote = null
-    ) {
-
-        return parent::isAvailable($quote);
-    }
-
     /**
      * Assign data to info model instance
      *
@@ -141,6 +81,8 @@ class BankSlip extends \Magento\Payment\Model\Method\AbstractMethod
         $info->setAdditionalInformation('installments', 1);
         $info->save();
 
+        parent::assignData($data);
+
         return $this;
     }
 
@@ -150,47 +92,5 @@ class BankSlip extends \Magento\Payment\Model\Method\AbstractMethod
     protected function getPaymentMethodCode()
     {
         return PaymentMethod::BANK_SLIP;
-    }
-
-    public function authorize(\Magento\Payment\Model\InfoInterface $payment, $amount)
-    {
-        /** @var Order $order */
-        $order = $payment->getOrder();
-        $customerId = $this->customer->findOrCreate($order);
-        $productList = $this->product->findOrCreateProducts($order);
-
-        $body = [
-            'customer_id' => $customerId,
-            'payment_method_code' => $this->getPaymentMethodCode(),
-            'bill_items' => $productList
-        ];
-
-        if ($installments = $payment->getAdditionalInformation('installments')) {
-            $body['installments'] = (int)$installments;
-        }
-
-        if ($bill = $this->bill->create($body)) {
-            if ($bill['charges'][0]['payment_method']['code'] === PaymentMethod::BANK_SLIP) {
-                $payment->setAdditionalInformation('print_url', $bill['charges'][0]['print_url']);
-                $payment->setAdditionalInformation('due_at', $bill['charges'][0]['due_at']);
-            }
-
-            if ($bill['charges'][0]['payment_method']['code'] === PaymentMethod::BANK_SLIP
-                || $bill['charges'][0]['payment_method']['code'] === PaymentMethod::DEBIT_CARD
-                || $bill['status'] === Bill::PAID_STATUS
-                || $bill['status'] === Bill::REVIEW_STATUS
-            ) {
-                $order->setVindiBillId($bill['id']);
-                return $bill['id'];
-            }
-            $this->bill->delete($bill['id']);
-        }
-
-        $this->psrLogger->error(__(sprintf('Error on order payment %d.', $order->getId())));
-        $message = __('There has been a payment confirmation error. Verify data and try again');
-        $order->setState(Order::STATE_CANCELED)
-            ->setStatus($order->getConfig()->getStateDefaultStatus(Order::STATE_CANCELED))
-            ->addStatusHistoryComment($message->getText());
-        throw new \Magento\Framework\Exception\LocalizedException($message);
     }
 }
