+++ etc/adminhtml/system.xml
@@ -52,11 +52,6 @@
                        type="text">
                     <label>Title</label>
                 </field>
-                <field id="order_status" showInDefault="1" showInStore="1" showInWebsite="1" sortOrder="100"
-                       translate="label" type="select">
-                    <label>New Order Status</label>
-                    <source_model>Magento\Sales\Model\Config\Source\Order\Status</source_model>
-                </field>
             </group>
         </section>
