+++ etc/config.xml
@@ -9,7 +9,7 @@
                 <order_status>pending</order_status>
                 <title>Vindi</title>
                 <allowspecific>0</allowspecific>
-                <payment_action>authorize_capture</payment_action>
+                <payment_action>authorize</payment_action>
                 <group>offline</group>
             </vindi>
             <vindi_bankslip>
