+++ i18n/pt_BR.csv
@@ -71,3 +71,4 @@
 "Vindi - Credit Card","Vindi - Cartão de Crédito"
 "Print Url: %s","Imprimir: %s"
 "Print your Bankslip","Imprimir Boleto"
+"The bank slip will be sent to your e-mail.","O boleto bancário será enviado para o seu e-mail."
