+++ view/frontend/layout/checkout_index_index.xml
@@ -1,6 +1,9 @@
 <?xml version="1.0" ?>
 <page layout="1column"
     xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:noNamespaceSchemaLocation="urn:magento:framework:View/Layout/etc/page_configuration.xsd">
+    <head>
+        <css src="Vindi_Payment::css/vindi.css" />
+    </head>
     <body>
         <referenceBlock name="checkout.root">
             <arguments>
