+++ view/frontend/web/template/payment/vindi-bankslip.html
@@ -21,6 +21,7 @@
             <!-- ko template: getTemplate() --><!-- /ko -->
             <!--/ko-->
         </div>
+        <div class="block-payment-bankslip"><!-- ko i18n: 'The bank slip will be sent to your e-mail.' --><!-- /ko --></div>
         <div class="checkout-agreements-block">
             <!-- ko foreach: $parent.getRegion('before-place-order') -->
             <!-- ko template: getTemplate() --><!-- /ko -->
